# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CTM API Docs",
    "version": "1.0.1",
    "description": "CTM all endpoints"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/user/by-wallet-address/{walletAddress}": {
      "get": {
        "tags": [
          "Open"
        ],
        "description": "Endpoint to fetch user basic information\\'s, User don\\'t have to be logged in",
        "parameters": [
          {
            "name": "walletAddress",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response body at code 200"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/revenue": {
      "get": {
        "tags": [
          "User",
          "Label"
        ],
        "description": "Endpoint to fetch user revenue information, User have to be logged in",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Expected response body at code 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revenue": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 6
                          },
                          "issuer": {
                            "type": "string",
                            "example": "did:ethr:0xD763016000cF8CA9594aEA52c351b30235A016F9"
                          },
                          "month": {
                            "type": "number",
                            "example": 7
                          },
                          "year": {
                            "type": "number",
                            "example": 2023
                          },
                          "rentRevenue": {
                            "type": "number",
                            "example": 0
                          },
                          "sellRevenue": {
                            "type": "number",
                            "example": 0
                          },
                          "totalRevenue": {
                            "type": "number",
                            "example": 0
                          },
                          "amountRented": {
                            "type": "number",
                            "example": 3
                          },
                          "amountSold": {
                            "type": "number",
                            "example": 0
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revenue": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 6
                          },
                          "issuer": {
                            "type": "string",
                            "example": "did:ethr:0xD763016000cF8CA9594aEA52c351b30235A016F9"
                          },
                          "month": {
                            "type": "number",
                            "example": 7
                          },
                          "year": {
                            "type": "number",
                            "example": 2023
                          },
                          "rentRevenue": {
                            "type": "number",
                            "example": 0
                          },
                          "sellRevenue": {
                            "type": "number",
                            "example": 0
                          },
                          "totalRevenue": {
                            "type": "number",
                            "example": 0
                          },
                          "amountRented": {
                            "type": "number",
                            "example": 3
                          },
                          "amountSold": {
                            "type": "number",
                            "example": 0
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not logged in"
          },
          "403": {
            "description": "Forbidden - wrong role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User",
          "Moderator",
          "SuperAdmin",
          "Label"
        ],
        "description": "to login provide DID token in bearer",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Expected response body at code 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "success": {
                      "type": "string",
                      "example": "true"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "success": {
                      "type": "string",
                      "example": "true"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/user/logout": {
      "post": {
        "tags": [
          "User",
          "Moderator",
          "SuperAdmin",
          "Label"
        ],
        "description": "Endpoint to fetch user basic information\\'s, User don\\'t have to be logged in",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Expected response body at code 200"
          },
          "401": {
            "description": "Unauthorized - User is not logged in"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/user/banned-users": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized - User is not logged in"
          },
          "403": {
            "description": "Forbidden - wrong role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}